New Installer class, makes installing assets available outside console command
