+++ Command/ThemesInstallCommand.php
@@ -49,11 +49,8 @@ class ThemesInstallCommand extends ContainerAwareCommand
         $activeTheme = $this->getContainer()->get('liip_theme.active_theme');
         $availableThemes = $activeTheme->getThemes();
 
-        $filesystem = $this->getContainer()->get('filesystem');
-
         $themesAssetsDir = $targetArg . DIRECTORY_SEPARATOR . "themes";
-        $filesystem->remove($themesAssetsDir);
-        $filesystem->mkdir($themesAssetsDir, 0777);
+
 
         // Do we hard copy or symlink ?
         $symlink = $input->getOption('symlink');
@@ -66,56 +63,8 @@ class ThemesInstallCommand extends ContainerAwareCommand
         foreach($availableThemes as $theme) {
             $output->writeln(sprintf('Installing assets for <comment>%s</comment>', $theme));
 
-            // Find theme path.
-            $themeLocator = $this->getContainer()->get("liip_theme.theme_locator");
-
-            // Search in bundles first.
-            $bundle = $themeLocator->locateThemeInBundles($theme);
-            if(!empty($bundle)) {
-                // Prepare the directory for this bundle.
-                $themesAssetsBundleDir = $themesAssetsDir . DIRECTORY_SEPARATOR . strtolower($bundle["bundle"]->getName());
-                if(!is_dir($themesAssetsBundleDir)) {
-                    $filesystem->mkdir($themesAssetsBundleDir, 0777);
-                }
-
-                // Found theme in bundle.
-                $originDir = $bundle["path"];
-                $targetDir = $themesAssetsBundleDir . DIRECTORY_SEPARATOR . $theme;
-
-                $output->writeln(sprintf('Found theme <comment>%s</comment> in <comment>%s</comment> installing in <comment>%s</comment> ', $theme, $bundle["bundle"]->getName(), $targetDir));
-            } else {
-                // Search in app/
-                $path = $themeLocator->locateThemeInApp($theme);
-                if($path) {
-                    $originDir = $path;
-                    $targetDir = $appThemesDir . DIRECTORY_SEPARATOR . $theme;
-                }
-
-                $output->writeln(sprintf('Found theme <comment>%s</comment> in <comment>%s</comment> installing in <comment>%s</comment> ', $theme, $originDir, $targetDir));
-            }
-
-            if($originDir && $targetDir) {
-                if($symlink) {
-                    // Symlink.
-                    $filesystem->symlink($originDir, $targetDir, true);
-                } else {
-                    // Hard copy.
-                    $this->hardCopy($originDir, $targetDir);
-                }
-            }
+            // Install assets for this theme.
+            $this->getContainer()->get('liip_theme.installer')->installAssets($theme, $themesAssetsDir);
         }
-    }
-
-    /**
-     * @param string $originDir
-     * @param string $targetDir
-     */
-    private function hardCopy($originDir, $targetDir)
-    {
-        $filesystem = $this->getContainer()->get('filesystem');
-
-        $filesystem->mkdir($targetDir, 0777);
-        // We use a custom iterator to ignore VCS files
-        $filesystem->mirror($originDir, $targetDir, Finder::create()->ignoreDotFiles(false)->in($originDir));
     }
 }
