+++ Resources/config/templating.xml
@@ -8,6 +8,7 @@
         <parameter key="liip_theme.file_locator.class">Liip\ThemeBundle\Locator\FileLocator</parameter>
         <parameter key="liip_theme.filesystem_loader.class">Liip\ThemeBundle\Twig\Loader\FilesystemLoader</parameter>
         <parameter key="liip_theme.theme_locator.class">Liip\ThemeBundle\Locator\ThemeLocator</parameter>
+        <parameter key="liip_theme.installer.class">Liip\ThemeBundle\Installer</parameter>
         <parameter key="liip_theme.active_theme.class">Liip\ThemeBundle\ActiveTheme</parameter>
         <parameter key="liip_theme.cache_warmer.class">Liip\ThemeBundle\CacheWarmer\TemplatePathsCacheWarmer</parameter>
         <parameter key="liip_theme.theme_auto_detect.class">Liip\ThemeBundle\Helper\DeviceDetection</parameter>
@@ -39,6 +40,11 @@
             <argument type="service" id="kernel" />
             <argument>%kernel.root_dir%</argument>
             <argument>%liip_theme.path_patterns%</argument>
+        </service>
+
+        <service id="liip_theme.installer" class="%liip_theme.installer.class%">
+            <argument type="service" id="liip_theme.theme_locator" />
+            <argument type="service" id="filesystem" />
         </service>
 
         <service id="liip_theme.theme_auto_detect" class="%liip_theme.theme_auto_detect.class%" />
